feat: expose label parsing feature to users in kili.labels
