+++ tests/services/label_data_parsing/parsing/test_resp_parsing.py
@@ -12,8 +12,8 @@ from kili.services.label_data_parsing.category import Category, CategoryList
 from kili.services.label_data_parsing.exceptions import FrameIndexError
 from kili.services.label_data_parsing.job_response import JobPayload
 from kili.services.label_data_parsing.json_response import ParsedJobs
-from kili.services.label_data_parsing.label import ParsedLabel
 from kili.services.label_data_parsing.types import Project
+from kili.utils.labels.label_parsing import ParsedLabel
 
 
 def test_attribute_category():
