+++ tests/services/label_data_parsing/test_label_utils.py
@@ -4,7 +4,7 @@ import pytest
 
 from kili.services.label_data_parsing.annotation import Annotation, AnnotationList
 from kili.services.label_data_parsing.category import Category, CategoryList
-from kili.utils.labels.parse_labels import parse_labels
+from kili.utils.labels.label_parsing import parse_labels
 
 
 def test_parse_labels_classification():
